+++ magellanx-core/src/main/java/com/ryanmoelter/magellanx/core/lifecycle/LifecycleAware.kt
@@ -5,9 +5,13 @@ public interface LifecycleAware {
 
   public fun show() {}
 
+  public fun start() {}
+
   public fun resume() {}
 
   public fun pause() {}
+
+  public fun stop() {}
 
   public fun hide() {}