+++ magellanx-sample-app/src/main/java/com/ryanmoelter/magellanx/doggos/MainActivity.kt
@@ -6,7 +6,6 @@ import com.ryanmoelter.magellanx.compose.setContentNavigable
 import com.ryanmoelter.magellanx.doggos.home.RootJourney
 
 class MainActivity : ComponentActivity() {
-
   lateinit var rootJourney: RootJourney
 
   override fun onCreate(savedInstanceState: Bundle?) {