+++ magellanx-sample-app/src/main/java/com/ryanmoelter/magellanx/doggos/randomimages/BreedImagesJourney.kt
@@ -3,11 +3,9 @@ package com.ryanmoelter.magellanx.doggos.randomimages
 import com.ryanmoelter.magellanx.compose.ComposeJourney
 
 class BreedImagesJourney : ComposeJourney() {
-
   override fun onCreate() {
     navigator.goTo(
-      ChooseBreedStep(chooseBreed = { breed -> navigator.goTo(DoggoImageStep(breed)) })
+      ChooseBreedStep(chooseBreed = { breed -> navigator.goTo(DoggoImageStep(breed)) }),
     )
   }
 }
-